created warrior header and correct dexterity typing in Ranger.h to float

+++ Ranger.h
@@ -15,7 +15,7 @@ class Ranger: public Player{
     void useUltimateSkill(Action* entity) override;
     ~Ranger();
     protected:
-    int dexterity;
+    float dexterity;
     
 
 };
